Rename mod config -> hardcode (#68)
So it doesn't interfere with the actual config subsystem of the
application
