+++ src/hardcode.rs
