+++ src/main.rs
@@ -19,7 +19,7 @@ mod looper;
 mod traits;
 mod midi;
 mod graphics_primitives;
-mod config;
+mod hardcode;
 mod measure;
 mod ui;
 mod state;
@@ -28,7 +28,7 @@ use midi::{AbsMidiEvent, PortMidiNoteTracker};
 use ui::Popup;
 use state::*;
 
-use config::*;
+use hardcode::*;
 
 fn print_devices(pm: &pm::PortMidi) {
     for dev in pm.devices().unwrap() {
